Refactored Oclc extract code into its own processor.  Added regex_split to PsuTraject module methods.
